requirements.txt
jsonschema==4.22.0
selene==2.0.0rc9
allure-pytest==2.13.5
python-dotenv==1.0.1
pydantic==2.5.2
requests==2.32.1
pytest==8.2.1
selene>=2.0.0rc4
Appium-Python-Client==4.0.0
